fix(motion-matching): rebuild perstep relocation against current main (#4044)
PR #4050 became stuck at conflicts because main moved 70+ commits forward
while it was open — including changes to the very files PR #4050 wanted to
move (#4042 added control_names imports, #4045 added --cost-mode arg).

Rebuilt cleanly:
- git mv MachineLearning/{train_dynamics_surrogate,optimize_torque_sequence_for_club,extract_dynamics_dataset}.py
  to src/shared/python/motion_matching/surrogate/perstep/{train,optimize,extract_dataset}.py
- replaced old paths with deprecation-warning shims that re-export from
  the new path (identity-equal class objects, verified by tests)
- copied the perstep/__init__.py from the original PR (exposes
  DynamicsMLP, TrainConfig, train_dynamics_surrogate, optimize_torque_sequence,
  extract_dataset symbols)
- carried over tests/unit/motion_matching/test_surrogate_perstep_relocation.py

All 5 relocation tests pass standalone.

Closes #4044.
