+++ src/engines/Simscape_Multibody_Models/3D_Golf_Model/MachineLearning/train_dynamics_surrogate.py
@@ -1,359 +1,27 @@
-"""Train a PyTorch MLP that maps golf swing state/load samples to kinematics."""
+"""Backwards-compatible shim for the relocated per-step trainer.
+
+The implementation moved to
+``src.shared.python.motion_matching.surrogate.perstep.train`` per issue #4044.
+This shim re-exports the public symbols and emits a ``DeprecationWarning`` on
+import so existing scripts that ``from train_dynamics_surrogate import ...``
+keep working for one release while callers migrate.
+"""
 
 from __future__ import annotations
 
-import argparse
-import json
-import logging
-import math
-import random
-from dataclasses import asdict, dataclass
-from pathlib import Path
+import warnings
 
-import numpy as np
-import pyarrow.parquet as pq
+from src.shared.python.motion_matching.surrogate.perstep.train import *  # noqa: F401, F403
+from src.shared.python.motion_matching.surrogate.perstep.train import (  # noqa: F401
+    DynamicsMLP,
+    TrainConfig,
+    main,
+)
 
-try:
-    import torch
-    from torch import nn
-    from torch.utils.data import DataLoader, TensorDataset
-except OSError as exc:  # Broken local torch installs fail this way on import.
-    raise SystemExit(
-        "PyTorch failed to import. Use Python 3.12 with a valid Torch install, "
-        "or create a CUDA venv as documented in README.md."
-    ) from exc
-
-
-SCRIPT_DIR = Path(__file__).resolve().parent
-DEFAULT_DATASET = SCRIPT_DIR / "data" / "processed" / "golf_inverse_ready.parquet"
-DEFAULT_OUTPUT_DIR = SCRIPT_DIR / "runs" / "baseline_mlp"
-LOGGER = logging.getLogger(__name__)
-
-
-@dataclass
-class TrainConfig:
-    dataset: str
-    output_dir: str
-    epochs: int
-    batch_size: int
-    learning_rate: float
-    weight_decay: float
-    hidden_sizes: list[int]
-    validation_fraction: float
-    test_fraction: float
-    seed: int
-    device: str
-    use_amp: bool
-
-
-class DynamicsMLP(nn.Module):
-    def __init__(
-        self, input_dim: int, output_dim: int, hidden_sizes: list[int]
-    ) -> None:
-        super().__init__()
-        layers: list[nn.Module] = []
-        prev = input_dim
-        for hidden in hidden_sizes:
-            layers.append(nn.Linear(prev, hidden))
-            layers.append(nn.LayerNorm(hidden))
-            layers.append(nn.SiLU())
-            prev = hidden
-        layers.append(nn.Linear(prev, output_dim))
-        self.net = nn.Sequential(*layers)
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        return self.net(x)
-
-
-def _load_column_metadata(parquet_path: Path) -> tuple[list[str], list[str]]:
-    metadata = pq.ParquetFile(parquet_path).schema_arrow.metadata or {}
-    input_columns = json.loads(metadata[b"input_columns"].decode("utf-8"))
-    target_columns = json.loads(metadata[b"target_columns"].decode("utf-8"))
-    return input_columns, target_columns
-
-
-def _load_arrays(
-    parquet_path: Path,
-) -> tuple[np.ndarray, np.ndarray, list[str], list[str]]:
-    input_columns, target_columns = _load_column_metadata(parquet_path)
-    read_columns = list(dict.fromkeys([*input_columns, *target_columns]))
-    table = pq.read_table(parquet_path, columns=read_columns)
-    frame = table.to_pandas()
-
-    x = frame[input_columns].to_numpy(dtype=np.float32, copy=True)
-    y = frame[target_columns].to_numpy(dtype=np.float32, copy=True)
-
-    finite_mask = np.isfinite(x).all(axis=1) & np.isfinite(y).all(axis=1)
-    dropped = int((~finite_mask).sum())
-    if dropped:
-        LOGGER.info("Dropping %s rows with non-finite values", dropped)
-    return x[finite_mask], y[finite_mask], input_columns, target_columns
-
-
-def _standardize(
-    train: np.ndarray,
-    values: np.ndarray,
-    eps: float = 1e-6,
-) -> tuple[np.ndarray, np.ndarray, np.ndarray]:
-    mean = train.mean(axis=0, dtype=np.float64).astype(np.float32)
-    std = train.std(axis=0, dtype=np.float64).astype(np.float32)
-    std = np.maximum(std, eps)
-    return (values - mean) / std, mean, std
-
-
-def _make_splits(
-    n: int,
-    validation_fraction: float,
-    test_fraction: float,
-    seed: int,
-) -> tuple[np.ndarray, np.ndarray, np.ndarray]:
-    rng = np.random.default_rng(seed)
-    indices = rng.permutation(n)
-    n_test = int(n * test_fraction)
-    n_val = int(n * validation_fraction)
-    test_idx = indices[:n_test]
-    val_idx = indices[n_test : n_test + n_val]
-    train_idx = indices[n_test + n_val :]
-    return train_idx, val_idx, test_idx
-
-
-def _rmse_by_column(
-    pred: np.ndarray, target: np.ndarray, columns: list[str]
-) -> dict[str, float]:
-    rmse = np.sqrt(np.mean((pred - target) ** 2, axis=0))
-    return {column: float(value) for column, value in zip(columns, rmse, strict=True)}
-
-
-def _r2_by_column(
-    pred: np.ndarray, target: np.ndarray, columns: list[str]
-) -> dict[str, float]:
-    residual = np.sum((target - pred) ** 2, axis=0)
-    centered = target - np.mean(target, axis=0)
-    total = np.sum(centered**2, axis=0)
-    r2 = np.where(total > 0, 1.0 - residual / total, np.nan)
-    return {column: float(value) for column, value in zip(columns, r2, strict=True)}
-
-
-def compute_phase_stratified_metrics(
-    pred: np.ndarray,
-    target: np.ndarray,
-    time_values: np.ndarray,
-    phase_breakpoints: dict[str, float] | None = None,
-) -> dict[str, float]:
-    """Optional per-phase RMSE report.
-
-    Time is *not* a model input. ``time_values`` is metadata only, used
-    here to bucket evaluation rows into swing phases for residual reporting.
-    Returns an empty dict when ``time_values`` has no positive range.
-    """
-    from .surrogate_validation import evaluate_per_phase, phase_stratified_split
-
-    if time_values.size == 0:
-        return {}
-    if float(np.max(time_values)) <= float(np.min(time_values)):
-        return {}
-    masks = phase_stratified_split(
-        {"t": np.asarray(time_values)},
-        time_col="t",
-        phase_breakpoints=phase_breakpoints,
-    )
-    return evaluate_per_phase(pred, target, masks)
-
-
-def _normalized_rmse_by_column(
-    pred: np.ndarray,
-    target: np.ndarray,
-    columns: list[str],
-) -> dict[str, float]:
-    rmse = np.sqrt(np.mean((pred - target) ** 2, axis=0))
-    std = np.std(target, axis=0)
-    normalized = np.where(std > 0, rmse / std, np.nan)
-    return {
-        column: float(value) for column, value in zip(columns, normalized, strict=True)
-    }
-
-
-def train(config: TrainConfig) -> None:
-    random.seed(config.seed)
-    np.random.seed(config.seed)
-    torch.manual_seed(config.seed)
-
-    dataset_path = Path(config.dataset)
-    output_dir = Path(config.output_dir)
-    output_dir.mkdir(parents=True, exist_ok=True)
-
-    x_raw, y_raw, input_columns, target_columns = _load_arrays(dataset_path)
-    train_idx, val_idx, test_idx = _make_splits(
-        len(x_raw),
-        validation_fraction=config.validation_fraction,
-        test_fraction=config.test_fraction,
-        seed=config.seed,
-    )
-
-    x_train_raw = x_raw[train_idx]
-    y_train_raw = y_raw[train_idx]
-    x_scaled, x_mean, x_std = _standardize(x_train_raw, x_raw)
-    y_scaled, y_mean, y_std = _standardize(y_train_raw, y_raw)
-
-    train_ds = TensorDataset(
-        torch.from_numpy(x_scaled[train_idx]),
-        torch.from_numpy(y_scaled[train_idx]),
-    )
-    val_x = torch.from_numpy(x_scaled[val_idx])
-    val_y = torch.from_numpy(y_scaled[val_idx])
-    test_x = torch.from_numpy(x_scaled[test_idx])
-
-    device = torch.device(config.device)
-    pin_memory = device.type == "cuda"
-    loader = DataLoader(
-        train_ds,
-        batch_size=config.batch_size,
-        shuffle=True,
-        num_workers=0,
-        pin_memory=pin_memory,
-    )
-
-    model = DynamicsMLP(
-        input_dim=x_scaled.shape[1],
-        output_dim=y_scaled.shape[1],
-        hidden_sizes=config.hidden_sizes,
-    ).to(device)
-    optimizer = torch.optim.AdamW(
-        model.parameters(),
-        lr=config.learning_rate,
-        weight_decay=config.weight_decay,
-    )
-    loss_fn = nn.MSELoss()
-    scaler = torch.amp.GradScaler(
-        "cuda", enabled=config.use_amp and device.type == "cuda"
-    )
-
-    history: list[dict[str, float]] = []
-    best_val = math.inf
-    best_path = output_dir / "best_model.pt"
-
-    for epoch in range(1, config.epochs + 1):
-        model.train()
-        train_loss = 0.0
-        train_count = 0
-        for batch_x, batch_y in loader:
-            batch_x = batch_x.to(device, non_blocking=pin_memory)
-            batch_y = batch_y.to(device, non_blocking=pin_memory)
-
-            optimizer.zero_grad(set_to_none=True)
-            with torch.amp.autocast(
-                "cuda", enabled=config.use_amp and device.type == "cuda"
-            ):
-                loss = loss_fn(model(batch_x), batch_y)
-            scaler.scale(loss).backward()
-            scaler.step(optimizer)
-            scaler.update()
-
-            train_loss += float(loss.detach().cpu()) * len(batch_x)
-            train_count += len(batch_x)
-
-        model.eval()
-        with torch.no_grad():
-            val_pred = model(val_x.to(device)).cpu()
-            val_loss = float(loss_fn(val_pred, val_y).cpu())
-        train_loss /= train_count
-        history.append({"epoch": epoch, "train_loss": train_loss, "val_loss": val_loss})
-        LOGGER.info(
-            "epoch=%03d train_loss=%.6f val_loss=%.6f",
-            epoch,
-            train_loss,
-            val_loss,
-        )
-
-        if val_loss < best_val:
-            best_val = val_loss
-            torch.save(
-                {
-                    "model_state_dict": model.state_dict(),
-                    "input_columns": input_columns,
-                    "target_columns": target_columns,
-                    "x_mean": x_mean,
-                    "x_std": x_std,
-                    "y_mean": y_mean,
-                    "y_std": y_std,
-                    "config": asdict(config),
-                },
-                best_path,
-            )
-
-    checkpoint = torch.load(best_path, map_location=device, weights_only=False)
-    model.load_state_dict(checkpoint["model_state_dict"])
-    model.eval()
-    with torch.no_grad():
-        pred_scaled = model(test_x.to(device)).cpu().numpy()
-    pred = pred_scaled * y_std + y_mean
-    target = y_raw[test_idx]
-
-    metrics = {
-        "best_val_loss_scaled": best_val,
-        "test_rmse_mean_unscaled": float(np.sqrt(np.mean((pred - target) ** 2))),
-        "test_rmse_by_target_unscaled": _rmse_by_column(pred, target, target_columns),
-        "test_nrmse_by_target_std": _normalized_rmse_by_column(
-            pred, target, target_columns
-        ),
-        "test_r2_by_target": _r2_by_column(pred, target, target_columns),
-        "n_rows": int(len(x_raw)),
-        "n_train": int(len(train_idx)),
-        "n_val": int(len(val_idx)),
-        "n_test": int(len(test_idx)),
-        "input_dim": len(input_columns),
-        "target_dim": len(target_columns),
-        "device": str(device),
-    }
-    (output_dir / "history.json").write_text(
-        json.dumps(history, indent=2), encoding="utf-8"
-    )
-    (output_dir / "metrics.json").write_text(
-        json.dumps(metrics, indent=2), encoding="utf-8"
-    )
-    LOGGER.info("%s", json.dumps(metrics, indent=2))
-
-
-def parse_args() -> argparse.Namespace:
-    parser = argparse.ArgumentParser(description=__doc__)
-    parser.add_argument("--dataset", type=Path, default=DEFAULT_DATASET)
-    parser.add_argument("--output-dir", type=Path, default=DEFAULT_OUTPUT_DIR)
-    parser.add_argument("--epochs", type=int, default=100)
-    parser.add_argument("--batch-size", type=int, default=8192)
-    parser.add_argument("--learning-rate", type=float, default=1e-3)
-    parser.add_argument("--weight-decay", type=float, default=1e-5)
-    parser.add_argument(
-        "--hidden-sizes", type=int, nargs="+", default=[512, 512, 256, 256]
-    )
-    parser.add_argument("--validation-fraction", type=float, default=0.1)
-    parser.add_argument("--test-fraction", type=float, default=0.1)
-    parser.add_argument("--seed", type=int, default=20260505)
-    parser.add_argument(
-        "--device", default="cuda" if torch.cuda.is_available() else "cpu"
-    )
-    parser.add_argument("--no-amp", action="store_true")
-    return parser.parse_args()
-
-
-def main() -> None:
-    logging.basicConfig(level=logging.INFO, format="%(message)s")
-    args = parse_args()
-    config = TrainConfig(
-        dataset=str(args.dataset),
-        output_dir=str(args.output_dir),
-        epochs=args.epochs,
-        batch_size=args.batch_size,
-        learning_rate=args.learning_rate,
-        weight_decay=args.weight_decay,
-        hidden_sizes=args.hidden_sizes,
-        validation_fraction=args.validation_fraction,
-        test_fraction=args.test_fraction,
-        seed=args.seed,
-        device=args.device,
-        use_amp=not args.no_amp,
-    )
-    train(config)
-
-
-if __name__ == "__main__":
-    main()
+warnings.warn(
+    "MachineLearning.train_dynamics_surrogate moved to "
+    "src.shared.python.motion_matching.surrogate.perstep.train (issue #4044). "  # noqa: E501
+    "Update your imports; this shim will be removed in a future release.",
+    DeprecationWarning,
+    stacklevel=2,
+)
